Move mutable input pipeline
