+++ test/core/pipeline/features/pipeline_run.feature
@@ -20,6 +20,7 @@ Feature: Pipeline running
         | that has two branches that merge |
         | that has different combinations of branches that merge and do not merge |
         | that has two branches, one of which loops back |
+        | that has a component with mutable input |
 
     Scenario Outline: Running a bad Pipeline
         Given a pipeline <kind>
